+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [3.2.1] - 2022-04-04
 ### Changes
-- mongo_generic: Added standard error output for dump process.
+- mongo_generic: Added standard error output for dump process and call process_log_file function.
 - mongo_dump: Added error log file for any errors produced during dump.
 - mongo_export: Added error log file for any errors produced during export.
 - Documentation changes.
+
+### Add
+- process_log_file: Checks and processes the log file to standard out and mail.
 
 
 ## [3.2.0] - 2020-12-16